pyproject.toml
[project]
name = "weather-forecast-app"
version = "0.1.0"
description = ""
authors = [
    {name = "Polina",email = "polina.revk@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-asyncio (>=0.26.0,<0.27.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "timezonefinder (>=6.5.9,<7.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
package-mode = false
